OSD: Add the support of finding out the dots latency bigger than selected ones
Add the implementation of finding out the dots that the op latency is
bigger than the selected ones. Originally, we wanted to implement the
feature that when clicking on a dot, then it shows the dots with
different colors which shares the same object id. However, the default
implementation of TapTool or Lasso_tool will highlight the selected dot
after selecting on that. And this default behavior makes the color
highlighting idea hard to implement(I don't know any way how to
disable the highlighting after selecting the dot). This implemenation
shows a demo of concept to find out the dots with bigger op latency than
the clicked one on a different component and can be used in showing out
the dots with the same object id in the future.

Signed-off-by: Gavin Guo <gavin.guo@canonical.com>